Fix for bug preventing new module file creation
sw-module.lua.j2 contains code that assumes the `base` variable
has been supplied in the task being executed.

This means that the `else` in the template for a default base
is never executed and makes setting a base mandatory.

The change allows `base` not to be set, and a default used
in the template.